refactor(freeze-updates): extract context-dependency pause/resume
Pull pauseContextDependency and resumeContextDependency (the
get/set descriptor swap that freezes a context-dep node's
memoizedValue) into freeze/context-dependency.ts.

freeze-updates.ts: 478 -> 419 (-59). 13 freeze-updates specs green.
